image-balena: disable space check for btrfs
With filesystem level compression such as used by btrfs, it's not
possible to assess at build time if the hostapp content will fit in the
inactive partition, as the compression ratio isn't known until it's
written to disk.

Disable this check to allow images to be built anyway.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>