tests/connectivity: Run the proxy tests with the actual redsocks uid of the DUT
Not all the boards we support have the redsocks uid as 995 in their rootfs so let's
fetch the actual redsocks uid from the DUT before running the proxy tests and
update that in the docker-compose.yml. We do so because the REDSOCKS_UID value
isn't substituted in the compose if the variable, even if it is passed trough
the cli.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>