initrdscripts: Only unlock LUKS partitions on the OS drive
In the current state the cryptsetup initrd script tries to unlock all
LUKS volumes in the system using the TPM. This includes user-defined LUKS
volumes that, if present, fail to unlock and make the system unbootable.
We should also not touch user-defined volumes in the first place.

This patch modifies the cryptsetup script to only unlock LUKS volumes
that are on the OS drive (same block device as the EFI partition).

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>