common: signing: improve debug output on failure
If the signing server's response is anything other than successful, such
as with an authentication failure or bad request, the HTTP status code
and response are hidden due to the --silent flag passed to cURL.

Drop the stdio redirect to the output file along with the --silent flag,
and instead use the -o parameter to output the response to the
appropriate file on success. This allows the status code and response to
be shown in the logs upon failure.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>