flasher: ensure image is signed before enrollment
The flasher image enrolls the secure boot keys before rebooting into
secured user mode and creating the encrypted luks volumes on disk.

If the image is not signed, the key enrollment will fail, and the
flasher will enter a loop trying to enroll them and rebooting.

Instead, skip the key enrollment if the image is not signed, resulting
in a non secure-boot installation.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>