common: os-helpers-fs: fix get_dev_path_from_label w/ luks
get_dev_path_from_label() calls lsblk to get the name and label of a
disk, then filters the list using the label and returns a /dev path.

The name returned when using a luks encrypted partition is the
/dev/mapper name, rather than the kernel's device mapper name under
/dev/dm-*. When assembling a path under /dev using the luks name, the
path is invalid, and the by-state links aren't created.

This leads to the rootfs hook failing to find and mount the resin-rootA
partition.

Change the attribute retrieved using lsblk to kname to fix this.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>