networkmanager: Make iptables wait for xtables.lock
iptables takes a file lock at /run/xtables.lock. By default, if
the file is locked, iptables will fail with error. When that happens,
the iptables rules won't be configured, and the shared mode
(for which we use iptables) will not be setup properly.

Instead, pass "--wait 2", to block.

The included patch has to be removed once we upgrade to next
NetworkManager version as it is already included upstream.

Change-type: patch
Signed-off-by: Zahari Petkov <zahari@balena.io>