flasher: fix secure boot setup with enrolled keys
When refactoring secure boot setup, a logic mistake in the purpose and
use of SECUREBOOT_VAR meant that devices booting the flasher with keys already
enrolled would bail out with an incorrect message about secure boot not
being supported in firmware.

This variable is `00` on systems with secure boot support in firmware,
but not enabled and enforced, `01` on systems where secure boot is
enforced, and empty when secure boot is unsupported.

Change this conditional to bail out only when the variable is empty,
indicating that secure boot is unsupported.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>