flasher: handle user mode system w/out secure boot
When a user runs the flasher with secure boot enabled in `config.json`,
the public keys used to validate the bootloader are enrolled. If any
other bootloader signature fails to validate against this public key, it
won't be executed.

If the user attempts to run the balenaOS flasher on that system again
without first enabling the secure boot option, the flasher won't enroll
keys, but the installed system will be signed. This will result in a
secure boot enabled system without full-disk encryption.

Bail out in this case so the user must choose to explicitly opt-in to
secure boot for the new installation, and full-disk encryption along
with it. Otherwise, the user must reset the enrolled keys to install
without secure boot.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>