resin-init-flasher: Build time check on INTERNAL_DEVICE_KERNEL only for flasher device types
Now that the installer script is added to the initramfs and is included
both for flasher and non-flasher device types, limit the
`INTERNAL_DEVICE_KERNEL` existence check to flasher images as
it is only used in the installer script.

Also, make the resin-init-flasher script check for it's existance before using
it.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>