common: remove systemd-udev-settle dependency from services
Udev docs [0] state that using systemd-udev-settle.service is not
recommended, as hardware enumeration is necessarily async, and there's no
clear point at which all hardware has been detected and initialized,
especially with the existence of hotplug.

Additionally, adding this service as a requirement results in the boot
process waiting for all queued events to be processed, even when they're
not relevant to the waiting service. This adds unnecessary delays to the
boot.

Remove this dependency completely from balenaOS to speed up the boot
process.

[0] https://www.freedesktop.org/software/systemd/man/systemd-udev-settle.service.html

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>