balena-keys: do not ship certificates in DER format
We have added DER certificates because it is the only format
that QEMU/TianoCore accepts in UEFI setup. Since we made setup mode
work properly, there is no need to program the QEMU keys manually
and therefore there is no reason to pollute the boot partition with
the DER files. If indeed necessary, they can always be extracted
from the ESL, which is still shipped. Shipping the DER files is also
not possible when db uses hashes instead of certificates, which we
want to move towards.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>