flasher: fix installation when in user mode w/ sb disabled
Previously, we bailed out of the installer when the system was in user
mode (keys enrolled) but the user had not opted in to secure boot, as it
was ambiguous whether the user actually wanted SB/FDE.

However, some systems come with vendor keys pre-enrolled, and a user may
simply turn off secure boot in the firmware setup menu without erasing
the keys. This would result in the installer bailing out even though
secure boot is disabled in the firmware menu.

Check that secure boot is enabled in addition to having keys enrolled
before bailing out.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>