Use hashes instead of certificates for secure boot image validation
This patch changes the validation of bootable images from certificate
signatures to a list of allowed hashes of binaries. This only applies
on db level, PK and KEK are still certificates.

The motivation is that certificates expire and we need to be sure
that even devices that have been lying on a shelf for several years
or whose CMOS battery has died and reset date to 1970-01-01 are still
bootable. Using hashes is more aligned with this use-case and also
more similar to the approach that embedded SoCs use.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>