Remove a bad check in Internet connection sharing test
In the Internet connection sharing test one of the checks may run
into a racing problem. The following command is holding the iptables
lock for 3 seconds while NetworkManager activates a connection with
sharing enabled:

`flock /run/xtables.lock sleep 3 & nmcli c up dummy & wait`

NetworkManager waits for 2 seconds for the lock to be released and
those three seconds should be enough for one of the iptables rules
to fail. However there is no guarantee that NetworkManager will start
adding the iptables rules that quickly - it may start adding those
after one out of those three seconds already passed, which will
lead to all iptables rules to be set at the end.

This check is non-essential for the test itself, so it is removed with
this commit.

Change-type: patch
Signed-off-by: Zahari Petkov <zahari@balena.io>