resin-u-boot.bbclass: Replace static patch resin-specific-env-integration-kconfig.patch
Because we use this patch with various u-boot versions it often happens that this patch
does not apply so we then need to rework it in the device integration layer. Instead it
would be better to have some code at configure time parsing the same file and inserting
the balena env dynamically, so regardless of u-boot versions we use.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>