resin-u-boot.bbclass: Default to u-boot Kconfig support
It should be safe to assume that boards now use newer u-boot versions
that all have Kconfig support so we default to that. This allows for
device repos not to specify it and use Kconfig support or if for some
reason there are boards with old u-boot versions they can overwrite the
UBOOT_KCONFIG_SUPPORT variable to 0.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>