os-helpers-fs: regenerate_uuid: skip remounting
Remounting filesystems is done on systems with a broken clock in order
to prevent tune2fs from bailing out when the last mounted time is in the
future. This resets the last mounted time to now.

However, the filesystem is immediately unmounted again without being
utilized, and the mount and unmount process is time consuming. Instead,
use `-e continue` to tell tune2fs to continue after an error, which
achieves the same result with less time and complexity.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>