os-helpers: fs: replace wait4file w/ inotifywait
Previously, scripts that required a file that is created asynchronously,
such as the by-state links created by a udev rule, would busy loop until
the file existed using wait4file(), part of os-helpers-fs.

This is time consuming and inefficient, as the minimum wait time adds
latency to the following operations. Instead, use inotifywait to receive
an event from the kernel when the relevant file or path is modified to
prevent unnecessary resource usage and busy waiting.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>