Update db and dbx hashes during HUP when secure boot is enabled
After moving to hashes for authenticating the allowed OS list, we need
to update the db variable on each HUP to make sure the new OS
will be bootable. After confirming that the update went through,
we need to update the dbx variable to make sure the old OS is
no longer bootable.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>