common: initrdscripts: run prepare early
The resin-update-state udev rules file will need to regenerate UUIDs on
boot for balenaOS filesystems without unique UUIDs. When mounting root,
a regenerated UUID is written to /run to override the bootparam so the
kernel is able to find the new UUID.

Run the prepare hook before udev to ensure the tmpfs is mounted at /run,
where this override is written to.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>