Ship separate GRUB images for secure boot
Shipping a single image with signature checks enabled will enforce
the signatures on non-secure-boot systems as well. GRUB does not have
a simple method to check whether secure boot is enabled that could
be embedded with the default built-in config.

With this patch we build two separate images - one enforcing
the signatures and the other one not, keeping the original behavior.
HUP and flasher both can detect if secure boot is enabled so they
put the correct image in place when installing/updating GRUB.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>