balena-config-vars: introduce unsafe fatrw copy
If a safe copy is preferred but non-critical, the unsafe fatrw command
can be used and if fatrw does not have enough resources to make a safe
copy it will fallback to a standard cp.

This is useful when performing hostOS updates for example where a safe
copy would only work for files that are read by fatrw also and that is
not the case for boot files for example.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>