resin-init-flasher: check that commands exist before calling
When running in the initramfs, the resin-device-progress package is not
installed as we cannot guarantee that the initramfs would be able to
bring up all types of network interfaces.

This commit adds a check for the script to exists instead of getting a
`command not found` when an API endpoint is defined.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>