initrdscripts: give the root device a chance to come up before cryptsetup
At this moment if the cryptsetup initrd script executes before the root
device comes up (e.g. when booting off a USB disk), it assumes
that encryption is not in place because it was not able to find
the EFI partition.

This patch adds a wait4file loop to the script that waits
for the /dev/disk/by-state directory. This is not tied to any particular
partition or device but since the directory does not exist by default
and is only created by a custom balenaOS udev rule, its existence
implies that the rule fired and a device with balenaOS partitions
is present in the system.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>