kernel-balena: add signing key to modules input chain
We have seen cases when using state cache with multiple parallel builds
were the kernel modules in the state cache are unchanged even though a
new kernel has been rebuilt.

The hypothesis that this tries to solve is that there is a window between
a new kernel build and a rebuild of the corresponding kernel modules where
the new kernel is used with the old state cached modules as the input chain
has not changed.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>