resin-init-flasher: reboot into signed flasher when provisioning secure boot
When programming keys for secure boot, some devices do not clear
the setup mode flag after a new PK is installed. In this case
flasher will reboot in order to ensure the keys are actually saved
and the device comes back with secure boot enabled. Since we changed
flasher to be unsigned by default, this reboot causes a security
violation.

With this patch flasher will add a new boot entry before issuing
the reboot so that signed flasher comes up and the installation process
can continue.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>