recipes-bsp/u-boot: Add script functions for checking kernel and dtb validity
The balena_kernel_load_crc_save / balena_fdt_load_crc_check functions
can be used in the device repository right after the kernel or
dtb is loaded and prior to calling the main boot function,
to verify if the memory areas of the kernel or fdt overlap
or may otherwise be re-written. These functions will save
create a file named overlap_detected in the boot partition,
which can be checked during automated testing. These checks
should not prevent booting in any way.

Signed-off-by: Alexandru Costache <alexandru@balena.io>