docs: Recommend BIOS password and no F-key shortcuts with secure boot
We have seen devices that won't change PCR1 hash when a temporary boot
order override was applied or secure boot was disabled via BIOS setup.
The implementation of what PCR1 actually measures is very
device-specific, but many of the risks can be mitigated by setting up
a BIOS password and disabling F-key shortcuts for interacting with
the firmware.

[skip ci]

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>