resin-init-flasher: Allow background device registration
314047e and b5c5214 made flasher block until the resin-device-register
service exits and made resin-device-register give up after 6 seconds
not to block infinitely when no network is available. This effectively
means that if the device fails to register within first 6 seconds,
it will never retry, flasher will not report status to the dashboard
and the device will only register on first boot.

This patch changes the logic back to resin-device-register trying
in the background in an infinite loop and moves the "give the device
a chance to register" delay to flasher itself. It also extends
the wait to openvpn as flasher already does that and wants VPN to run
to be debuggable - in case flashing fails, it would be possible
to SSH in via the VPN and look at the logs remotely.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>