grub-efi: Use ephemeral GPG keys instead of signing service
When signing for secure boot, GRUB embeds a public GPG key that it uses
to verify a signature of each file it loads. At this moment it uses
a key stored in the signing service so each GRUB build uses the same
key which makes GRUB unlikely to change between two OS builds. This
is problematic because with secure boot, we would like a separate GRUB
binary for each OS version to be able to properly blacklist previous
OS versions.

This patch removes the dependency on the signing service and makes
GRUB use a locally generated GPG key that changes with each build.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>