os-helpers-fs: regenerate_uuid: redirect subcommands
The regenerate_uuid() function is used to generate unique UUIDs for
balenaOS partitions on boot. When regenerating UUIDs inside the
initramfs, we need the newly generated root UUID to override the root
bootparam in /run.

Redirect unrelated subcommand output to stderr so that we can save the
stdout from regenerate_uuid() without parsing for the UUID.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>