Make Engine healthcheck performance test aware of device type
We previously tried to use a single time limit for the execution of the
healthcheck test on all device types. This was causing occasional false
positives in our Continuous Integration pipeline, though -- especially
on slow devices like Pi Zeros and the generic-aarch64, which runs on
emulated hardware.

This commit addresses this issue, this commit:

1. Limits execution for device types for which we have collected enough
   data to have a good idea of how long the test should take.
2. Uses time limits specific for each device type.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch