initramfs: stub wait from fsuuidsinit hook
In a previous commit, we moved the generation of unique filesystem UUIDs
into a udev rule, which is performed asynchronously, and replaces the
old fsuuidsinit hook.

Part of the functionality of the old hook was waiting for the by-state
links to be created. Without this, scripts that expect these links to
exist, such as resindataexpander, may fail to find partitions, and
silently fail to perform their intended tasks.

Reimplement this wait in the same order as before to ensure these
existing scripts continue to work as expected. Once these hooks are
moved to udev rules or made more robust, we can remove this stub.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>