grub-conf: Do not hardcode the path for grub_extraenv
At this moment grub.cfg sources /grub/grub_extraenv which works fine
on MBR systems, however on EFI systems this does not work because GRUB
is installed in /EFI/BOOT/ rather than /grub/.

This patch replaces the hardcoded /grub with ${prefix} which should
expand to the appropriate directory regardless of the platform. It also
removes the loading of grub_extraenv from the secure boot variant
of the GRUB config since this would not load without a signature anyway.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>