initrdscripts: add a script that switches between rootA and rootB
This is done by the bootloader (uboot/grub) at this moment but as we
are moving towards the balena 2nd stage bootloader, it needs to be
moved into the initramfs.

This adds a standalone recipe - by default yocto tries to build all
modules defined in the initramfs-framework recipe, which breaks
on armv7 when abroot is defined there. This is because it depends
on grub-editenv which is not supported on armv7.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>