os-helpers-fs: fix returning multiple matching devices
When searching for devices matching the glob list in
get_internal_device(), a glob match breaks from a nested loop rather
than the parent loop, allowing the function to output multiple matches.

When running the flasher, this results in the script failing with an
incorrect path to the internal disk.

Break from the parent loop to fix this behavior.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>