update-balena-supervisor: ensure target supervisor is running
If the target supervisor image is already cached but there is no
container running with it, the update script would just exit without
actually running the target supervisor.

This commit checks whether there is a running container using the
target image and restarts the supervisor if there is none.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>