resindataexpander: authenticate cryptsetup-resize
According to the man page for cryptsetup-resize:

  If cryptsetup detected volume key for active device loaded in
  kernel keyring service, resize action would first try to retrieve the
  key using a token. Only if it failed, it’d ask for a passphrase to
  unlock a keyslot (LUKS) or to derive a volume key again (plain mode).
  The kernel keyring is used by default for LUKS2 devices.

In some cases, specifically when running a newly flashed secure boot
enabled system in QEMU, the cryptsetup-resize execution in
resindataexpander will request the LUKS passphrase. Presumably because
the utility was unable to retrieve the key from the kernel keyring,
possibly a race condition.

Pass the LUKS passphrase to cryptsetup-resize explicitly as a parameter
to avoid this.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>