os-helpers-tpm2: separate authentication from crypto
When encrypting the LUKS passphrase, we need the ability to construct a
policy that can logically OR together multiple policies, such as when
the machine may or may not measure binaries loaded through EFI boot
services into PCR7.

We also need the ability to update the sealing policy to revoke
previously valid configurations, such as after hostapp-healthcheck
completes successfully. Ideally, this should be completed before
modifying any efi variables, to prevent the system from becoming
unbootable in the event of an interrupted update.

These requirements necessitate the ability to create sealing policies
and authenticate against them outside of the hw_{en,de}crypt_passphrase
functions.

This commit allows the caller to setup the sealing policy when
encrypting, and choose what kind of authentication to use when
decrypting.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>