meta-balena-common: Move code from meta-balena-kirkstone
The code moved from meta-balena-kirkstone was not really specific to
kirkstone so let's move it here so that future branches for newer yocto
releases which we'll base off kirkstone don't continue to add this
unneeded  duplication.
There are other meta-balena-* directories that still contain the
duplication we moved from meta-balena-kirkstone but we're not really
concerned with that because going forward those old directories will
naturally get deprecated.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>