os-helpers-logging: replace broken container check
Detecting whether the script is running in a PID namespace by checking
`/proc/1/sched` does not work for newer kernels
(see https://github.com/systemd/systemd/pull/17917).

Properly detecting this is too cumbersome for a bash logging script,
see https://github.com/systemd/systemd/pull/17902, however, falling
back to the last check, that is, seeing if `/.dockerenv` exists is easy
enough and works for our use case.

This script will only be called from the hostOS, and the only case it is
called from a container is during HUP and the container is always a hostOS
image. So even though the interface chosen by moby, a file under /,
is a bad interface in general, it works fine for the specific limitations
of balenaOS.

Also, check for `/run/.containerenv` which is the equivalent interface
for podman for future proofing.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>