hostapp-update-hooks: add debug mode
This provides an easy switch to enable tracing on HUP hooks that works
both on old and new OS hooks as enabling it depends on a config.json
setting.

It is meant to debug field issues with HUP failure where all we see is:
```
Before hooks (old os) ran successfully
Failed to run the new hooks. Running current hooks..
```
For example:
https://jel.ly.fish/support-thread-1-0-0-front-cnv-eq6ipvx

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>