initrdscripts: move mounting of log mounts to finish module
The log files are kept in the /run partition that also houses
the udev database.

By moving the /run partition to $ROOTFS_DIR/run in the migrate
module, the udev database persists across reboots.

The udevcleanup module was introduced to cleanup the database
so it's regenerated on the final rootfs as the rules are
different between initramfs and final rootfs. By moving /run
before udevcleanup ran, the cleanup was not effective.

This commit moves the relocation of the /run mount to the
finish module, after udevcleanup is run, and groups it with
the other relocations needed for the pivot switch.

Looking at git history, the relocation of the /run partition
was initially done in the rootfs module, so when the migrate
module was introduced and needed to mount the rootfs, this
relocation was just moved into the migration module itself,
without considering later modules like finish.

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>