hostapp-update-hooks: generalize hooks for secure boot support
Allow the bootfiles script to install to configurable mountpoints so that
it can be used to install to more than one boot partition. This is useful
for example in secure boot systems that split the boot partition in two,
encrypted and non-encrypted.

Also, check that the non-encrypted partition exists in the rollback
hook.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>

SQ hostapp-update-hooks 1-bootfiles