initrdscripts: remove nr_cpus kernel arg before kexec
The kexec initrd script currently removes the maxcpus=0 kernel arg
which we use to put the system into non-SMP mode. This however does
not work on all platforms and nr_cpus=1 seems to be a more robust
solution, so with this patch nr_cpus will be removed as well.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>