resin_update_state_probe: ignore RAID members when looking for root
At this moment resin_update_state_probe looks for the root partition
by UUID and assumes that only a single device is returned. This
assumption breaks when the root is on a MD RAID1 device as not only
the virtual MD device holds a filesystem with the given UUID,
each member does as well.

This patch makes resin_update_state_probe ignore devices
that are RAID members which in that case will only return
the MD device as expected.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>