networkmanager: make FORWARD rules for shared interfaces last in chain
At this moment there is a race condition between NetworkManager
and the engine when a shared interface is configured. If the interface
is configured first and the engine second, the containers are allowed
to access DHCP hosts behind the shared interface. If the engine comes
up first and the shared interface second, access will be denied.

This patch adds a dispatcher script that always configures the firewall
rules as if the engine came up last. This does not really address
the underlying issue but it overcomes the race condition and makes
the behavior deterministic, which is good enough at this point.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>