balena-image-bootloader-initramfs: add modules needed for secure boot
The balena bootloader needs to mount encrypted disks to kexec the final
kernel which is stored in the encrypted root partitions.

It also needs to run the data partition expander twice on boot, once in the
balena bootloader that expands the disk, and later on the final
initramfs to expand the file system.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>