Include fsck module in balena bootloader initrd
The balena bootloader needs to mount rootA or rootB in order to find
the kernel it is supposed to kexec into. When the rootfs is dirty
and the bootloader mounts it without first running fsck, this causes
corruption on devices without an RTC (though only when last mount time
is in the future). A subsequent attempt to run fsck (in the balenaOS
initrd) will bail out and log a message asking to run fsck manually.

Including the fsck module in the bootloader initrd always runs fsck
before mounting and avoids the issue.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>