hostapp-update-hooks: Soft include balena-config-defaults
a203bcdfd567c0cc4b4ed9de493513142cd7463f introduced a dependency
on /usr/sbin/balena-config-defaults to hostapp-update-hooks, however
during HUP the script is not only executed in the "new" OS container
but directly in the context of the "old" OS as well, so
/usr/sbin/balena-config-defaults needs to exist there.
The file was introduced in balenaOS v2.99.28, so trying to HUP
from anything before that will fail.

This patch changes this to a soft dependency so even if the file
is missing HUP will continue.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>