systemd: update_state_probe: Use a file mutex to avoid race condition
As soon as the UUID is modified udev re-runs the rules for the partition.
However, the rule expects the new root UUID to be cached in a file, and
if the udev rule gets there before the file is created it fails.

This commit waits on a lock file mutex before accessing said file.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>