initrdscripts: fsuuidinit: use file based mutex to avoid race condition
As soon as the UUID is regenerated udev runs the correspondign rules.

However, the rules expect the new UUID to be cached in a file, so there
is a race condition between the creation of the file and the udev rule.

This commit avoid the race condition by using a file mutex that the
udev rule can wait on.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>