rollback-health: move apply-dbx to HUP commit hook
This operation is done after rollback-health completes and the new OS is
running to ensure the OS is healthy before appending to the forbidden
signatures list.

Move this out of rollback-health and into a HUP commit hook, which
allows it to be excluded from OS images that don't use EFI or support
secure boot.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>