os-helpers-sb: Silence command when checking for user_mode_enabled
is_secured will use "command" to check whether "user_mode_enabled"
is defined, which was added to prevent the error message
"user_mode_enabled: command not found" as this is only defined
on EFI systems. However now on EFI systems it echoes
"user_mode_enabled" each time the command is found which is also
not what we want.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>