NetworkManager: replace internal DHCP client by dhclient
Testing the behavior, not to be merged.

Als the last yocto version I was able to build this with is dunfell.
As dhclient is orphan upstream, the dhcp recipe has been removed
from recent poky versions. It won't build even after backporting
the recipe, because it depends on an old version of bind.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>