os-helpers-tpm2: firmware_measures_efibins: silence grep
The firmware_measures_efibins function outputs different strings
depending on whether the TPM event log is available, and whether or not
EFI binaries are measured into PCR 7 as indicated in the event log.

We don't need to print the output of the parsed event log, so redirect
it.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>