hup: signed-update: update boot files as needed
Unconditionally update the kernel and second stage bootloader when the
content on disk doesn't match the binaries shipped in the hostapp.

Previously this was only done when migrating, but the kernel, and
consequently the second stage bootloader, change every build. This means
firmwares which measure EFI binaries into PCR 7 won't boot unless the
second stage bootloader is updated to match the digests enrolled in the
security database.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>