rollback-health: mount securityfs in old_rootfs
When rollback-health runs, a failing healthcheck causes the
hostapp-update hooks to be run from the inactive partition, to make the
inactive system bootable again.

The 0-signed-update hook, which updates the sealing policy for secure
boot enabled systems, reads from the securityfs mounted at
/sys/kernel/security in order to parse the TPM event log.

If this filesystem isn't mounted, the hook will improperly detect that
the TPM event log isn't available, and unneccessarily create a combined
policy when a single PCR policy would suffice.

Mount this filesystem in old_rootfs before chrooting to fix this.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>