os-helpers-tpm2: fix empty efivar reads
When reading from efi variables in hostapp-update hooks during rollback,
tcgtool will improperly read zero bytes from efivar files. This results
in an improper calculation of the PCR 7 digest, and an unbootable
system.

Read the file contents, skipping the first four bytes that are
attributes, and pipe the data directly to tcgtool to work around this.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>