networkmanager: remove race condition from shared dispatcher script
The shared dispatcher script runs when a shared iptables rules is
identified and it moves it to the last place in the forward chain to
avoid race conditions with the engine rule addition.

This commit refactors the script to remove the possibility of a rule
changing positions between being identified by line number and removing
it.

The script now only fetches the rule line number once and then uses its
description to both append and remove.

It also has a final check to make sure the rule has been moved to the
last position of the forward chain, and tries a series of times if that
is not the case.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>