NetworkManager: do not use FW rule numbers in shared dispatcher script
Manipulating the firewall rules by index introduces a race condition.
Both NetworkManager and balenaEngine add the rules to the top
of the FORWARD chain instead of appending, so if we first look up
a rule by number and then use the number to refer to it, we can not
guarantee that the rule number has not changed (iow the rule has not
been moved down) in the meantime.

This patch removes the use of rule numbers completely and makes
the "shared" dispatcher script refer to the rules by definition.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>