rtl8192cu: Remove this unmaintained out-of-tree kernel driver
The source code for this driver hasn't been updated in 9 years so
let's remove it by also considering we were not using it. Instead,
machines that need to use a driver for this WiFi chipset can selectively
enable the in-tree kernel driver which has been improved in the last years.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>