plymouth: fix build error when disabling udev
When building plymouth 24.004.60 with GCC 13.2 and the `-Dudev=disabled`
configuration, the build fails with several undefined symbols.

    | /work/build/tmp/work/core2-64-poky-linux/plymouth/24.004.60/recipe-sysroot-native/usr/bin/x86_64-poky-linux/../../libexec/x86_64-poky-linux/gcc/x86_64-poky-linux/13.2.0/ld: src/libply-splash-core/libply-splash-core.so.5.0.0.p/ply-device-manager.c.o: in function `create_devices_for_terminal_and_renderer_type':
    | ply-device-manager.c:(.text+0xf1e): undefined reference to `add_input_devices_to_renderer'
    | collect2: error: ld returned 1 exit status

It's unclear what's causing this build error, but it appears to be a bug
in plymouth. Building with clang results in a similar undefined symbol
error.

Rather than clearing the entire PACKAGECONFIG variable, explicitly
remove unneeded configs, leaving udev enabled to fix the build.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>