hostapp-update-hooks: Re-add check for UEFI to signed-update hook
In 328222014146f0116e0208443f3e255d0e85ef15 we have removed
the signed-update hook from systems that do not have EFI
in MACHINE_FEATURES. This on its own makes sense, however together
with it we have also removed the runtime check for whether the running
system is actually booted in UEFI mode.

This effectively means it is no longer possible to update the host OS
on a device type able to boot in both UEFI and BIOS modes
(intel-nuc and genericx86-64-ext) when booted in BIOS mode,
as the signed-update hook is executed unconditionally and fails
if the device is not running UEFI.

This patch re-adds the runtime check to only execute the hook
if the system is actually booted in UEFI mode.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>